Deprecate the with statement in QML
It is generally considered deprecated in JavaScript and its use disables
a whole range of optimizations that we would like to apply in the future.

Therefore this patch will issue a warning if the with statement is detected.

This change is also documented, along with the plan on enabling strict mode
in the future.

Change-Id: Ie60f0536e0bdd6ecc537d8e34efbd8868bcad743
Reviewed-by: Lars Knoll <lars.knoll@digia.com>