Fix crashes in workerscript tests (qqmlecmascript and qquickworkerscript)
We need access to the QNam for a worker and other bits from the v8Engine,
so for that it's easiest right now to store the v8engine pointer directly.

There aren't many occurrences of v8engine left and we're gradually getting rid
of them, but this let's the tests not crash at least :). Once the transition
is complete, QV8Engine should collapse and v4->v8Engine->jsengine should become
v4->publicEngine again.

Change-Id: I98558dd687875f20f1dbe25381ee7efe502c7f24
Reviewed-by: Lars Knoll <lars.knoll@digia.com>