Fix two qqmlecmascript tests (assignSequenceTypes one of them)
In various places in the engine we do QVariant(v4Value.booleanValue()),
which unfortunately constructs a QVariant from an unsigned int.

IIRC we used Bool(uint) as type in the run-time to ensure that the compiler
allocates an entire general purpose register for JIT calls on all platforms, so
those functions remain untouched. But I think it's safe to at least change the
getters in Value, as those are never called directly from generated code.

Change-Id: Ia235141b8bcfc38403a7967436c68e73e7eee669
Reviewed-by: Lars Knoll <lars.knoll@digia.com>