move quick dialogs example into an artificial subdirectory again
The trouble with Id28d5718b4b7b475dcd7d62e31c4634219dc3ddc is that
the build assumes that all example directories are subdirectories, so
the examples-manifest.xml will point to the wrong place and creator
ends up not having the example on the welcome page.  That's more rigid
than it should be, but we might as well use the opportunity to
distinguish system dialogs from any QML-only dialogs that might be
added to this module later on.

Change-Id: I6e7261e096d75df560c6ce608af4c54b82e52d13
Reviewed-by: Topi Reiniö <topi.reinio@digia.com>
Reviewed-by: Caroline Chao <caroline.chao@digia.com>