Fix an obvious typo in the pool allocation sizing.
The pooled allocator in V4 preallocates the memory in a reasonably
sized chunks to limit the OS overhead of the allocations.
The chunks are supposed to be exponentially bigger, but the chunk sizes
should not depend on the size of the original allocation request.
Before the fix, the first allocation in every bucket basically
preallocated 64k slots of given size instead of 64k bytes, thus even
a trivial application ended up with 28MB of such heap.
After the fix, the same application is happy with <400kB
of this managed heap.

Change-Id: I1deae95bb6f45b8c67afe71d3a560b787357b31b
Reviewed-by: Simon Hausmann <simon.hausmann@digia.com>