Fix JavaScript stack back trace generation in release builds on Windows
We need to be able to reliably generate stack back traces for JavaScript
calls, and since we use the C stack we therefore need to be able to reliably
walk the stack.

On most platforms the stack back trace generation is tied to the same
mechanism that's used to unwind the stack during exception handling,
which uses compiler generated per-function tables that allow the run-time
to unwind the stack even when no regular stack frame was created.

Unfortunately on i386 on Windows there are no unwinding tables in use (as
opposed to all other architectures Windows runs on), and therefore we have no
reliable way of generating a stack back trace if functions can omit the
generation of a stack frame. Therefore we have to disable this compiler
optimization, using -Oy- for MSVC (see http://msdn.microsoft.com/en-us/library/2kxx5t2c(v=vs.71).aspx )
and -fno-omit-frame-pointer for gcc.

Technically this change needs to be done only in places where we support
throwing or catching V4 exception as well as code that is traversed during the
unwinding. Due to the use of internal V4 api throughout the entire module,
this patch is applied to .qmake.conf.

This also fixes tst_qqmlvaluetypes

Change-Id: I21a9e5522741446de25a5d0046f7e34f741f7722
Reviewed-by: Erik Verbruggen <erik.verbruggen@digia.com>