Fix crashes in tst_qquickcanvasitem with older Ubuntu toolchains
Older versions appear to be much more sensitive to missing
unwind tables, and as it turns out: We were compiling QtQuick (not QtQml)
with -fno-exceptions.

Since the entire module has access to the private V4 API and may therefore
be subject to exception traversal, make sure to compile the entire module
with CONFIG += exceptions.

Change-Id: If2d029879392c22af7da2f1bba94366d51518226
Reviewed-by: Erik Verbruggen <erik.verbruggen@digia.com>