Correctly update flickable visibleArea.heightRatio when geometry changes
Without this fix the visibleArea.heightRatio and widthRatio values
were only updated on geometry changes when flicking was active.

So when setting the flickable geometry to the content geometry and
thereby disabling flicking the ratios were not updated.

This could for example cause wrong scrollbar renderings.

The ratios are now also calculated directly after accessing the
visibleArea property for the first time.

The new autotest covers both problems.

Change-Id: I54ba606524557fb328a198c312c1f65eb125c5a3
Reviewed-by: Alan Alpert <aalpert@blackberry.com>