Add initial draft of the compiled data structures for JS and QML
To be able to cleanly separate the compilation stage from the runtime,
the compiler will create a read-only data structure, that is described
in this file.

This will later on also allow us to fully compile JS and QML in separate
threads, and longer term replace the qqmlcompiler/vme infrastructure.

Change-Id: Iafe6d527c0994c3bc17ba1630959ffd32cc74067
Reviewed-by: Simon Hausmann <simon.hausmann@digia.com>