Expose drag threshold in MouseArea
In several cases such as for creating a Slider, it is useful to  have a
0 pixel threshold in order to initiate a drag operation. We have
previously hardcoded this to a platform dependent (usually 10 pixel)
value for MouseArea.

Note that we have no way of indicating the version/revision number
in a grouped property for documentation at the moment.

In addition we deliberately had to remove the REVISION from the
property because it is blocked by QTBUG-33179. However, since
this is not a user-creatable type it should not cause any
issues in practice.

This patch adds MouseArea.drag.threshold in order to improve on this.

Change-Id: Ia4871e64fab39e30c4494f00be99ad38cdd630df
Reviewed-by: J-P Nurmi <jpnurmi@digia.com>