Explicitly call setCursor in the QQuickMouseArea constructor
QQuickItemPrivate has a boolean hasCursor that is used by QQuickWindow
to determine whether to use the cursor for an item. QQuickMouseArea has
a default cursorShape of Qt.ArrowCursor that is ignored due to hasCursor
not being set. Thus, call setCursor in the QQuickMouseArea constructor
to sync the cursor state of the QQuickItem and the MouseArea.

Task-number: QTBUG-33157

Change-Id: I56402564bd72bdc89d6ef8e5bc75bfd07c096202
Reviewed-by: Josh Faust <jfaust@suitabletech.com>
Reviewed-by: Alan Alpert <aalpert@blackberry.com>